requirements.txt
arrow==1.3.0
asgiref==3.8.1
binaryornot==0.4.4
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cloudinary==1.44.0
cookiecutter==2.6.0
crispy-tailwind==1.0.3
dj-database-url==2.3.0
Django==5.1.4
django-browser-reload==1.18.0
django-cloudinary-storage==0.3.0
django-crispy-forms==2.3
django-tailwind==4.0.1
grpcio==1.71.0
grpcio-tools==1.71.0
gunicorn==23.0.0
idna==3.10
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
packaging==25.0
pillow==11.0.0
protobuf==5.29.4
psycopg==3.2.6
psycopg-binary==3.2.6
psycopg2-binary==2.9.10
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
types-python-dateutil==2.9.0.20241206
typing_extensions==4.13.2
urllib3==2.4.0
whitenoise==6.9.0
